third_party/WebKit/LayoutTests/http/tests/devtools/sources/color-swatch-position-expected.txt

Tests that color swatch positions are updated properly.

Initial swatch positions:
Line 1, Column 11: blue
Line 1, Column 35: #ffffff
Line 4, Column 11: rgb(0, 0, 0)

Running: testEditSpectrum
Line 1, Column 11: #008000
Line 1, Column 38: #ffffff
Line 4, Column 11: rgb(0, 0, 0)

Running: testAddLine
Line 2, Column 11: #008000
Line 2, Column 38: #ffffff
Line 5, Column 11: rgb(0, 0, 0)

Running: testDeleteLine
Line 4, Column 11: rgb(0, 0, 0)

Running: testAddColor
Line 2, Column 7: hsl(300, 100%, 35%)
Line 4, Column 11: rgb(0, 0, 0)

Running: testInvalidateColor
Line 4, Column 11: rgb(0, 0, 0)

Running: testBookmarksAtLineStart
Line 4, Column 11: rgb(0, 0, 0)
Line 6, Column 0: #ff0
